On parie que vous n'en pusherez qu'un ?

+++ pathFollower.hpp
@@ -50,7 +50,7 @@ class PathFollower
         static void rotateCallback(struct motionElement* element);
 
         static bool isOutsideLand(int x, int y);
-        static std::pair<std::pair<double,double>,std::pair<double,double> > projectInLand(int x, int y, int prevX, int prevY, int& type, float& dist);
+        static std::pair<std::pair<double,double>,std::pair<double,double> > projectInLand(int x, int y, int prevX, int prevY, int& type, double& dist);
 
         static void resetPosition(const std::pair<double,double>& v);
         static std::pair<double,double> getCurrentPos();
@@ -74,7 +74,7 @@ class PathFollower
         static std::list<double> angles, distances;
         static std::list<bool> recalibrate;
         static std::list<std::pair<double,double> > positionAfterRecalibration;
-        static std::list<float> distancesRecalibration;
+        static std::list<double> distancesRecalibration;
         static std::list<int> type_recal;
 
         static std::pair<double,double> prevPosition;
